Move GetMemoryUsage to its own submodule
