+++ mempool/src/lib.rs
@@ -23,14 +23,15 @@ use common::primitives::{BlockHeight, Id};
 use common::time_getter::TimeGetter;
 pub use interface::MempoolInterface;
 
-use crate::config::GetMemoryUsage;
 use crate::error::Error as MempoolError;
 use crate::mempool_interface_impl::MempoolInterfaceImpl;
+use get_memory_usage::GetMemoryUsage;
 
 pub use crate::mempool_interface_impl::SystemUsageEstimator;
 
 mod config;
 pub mod error;
+mod get_memory_usage;
 mod interface;
 mod mempool_interface_impl;
 pub mod rpc;
