+++ mempool/src/mempool_interface_impl/pool.rs
@@ -45,6 +45,8 @@ use utils::tap_error_log::LogError;
 
 use crate::error::Error;
 use crate::error::TxValidationError;
+use crate::get_memory_usage::GetMemoryUsage;
+use crate::get_memory_usage::MemoryUsage;
 use crate::mempool_interface_impl::MempoolMethodCall;
 use crate::tx_accumulator::TransactionAccumulator;
 use crate::MempoolEvent;
